User

+++ controllers/form.js
@@ -1,29 +1,46 @@
 const formModel = require("../models").forms;
-const imageModel = require("../models").images;
 const fs = require("fs");
-
+//add me
 module.exports = {
-  addForm: async ({ INFORMATION }) => {
+  addForm: async ({ INFORMATION }, req) => {
     try {
-      INFORMATION = JSON.parse(JSON.stringify(INFORMATION));
-      console.log(INFORMATION);
-      const information = await formModel.create(INFORMATION);
-      console.log(information);
-      return information;
+      if (req.role !== null && req.role !== undefined) {
+        INFORMATION = JSON.parse(JSON.stringify(INFORMATION));
+        INFORMATION["date"] = new Date(INFORMATION.date)
+          .toISOString()
+          .replace(/T/, " ")
+          .replace(/\..+/, "");
+        INFORMATION["senderID"] = req.userID;
+        const information = await formModel.create(INFORMATION);
+        console.log(information);
+        return information;
+      }
     } catch (error) {
       throw error;
     }
   },
-  deleteForm: async ({ INFORMATION }) => {
+  getForm: async (args, req) => {
     try {
-      INFORMATION = JSON.parse(JSON.stringify(INFORMATION));
-      await formModel.deleteOne({
-        title: INFORMATION.title,
-        image: INFORMATION.image,
-      });
-      await imageModel.deleteOne({ img: INFORMATION.image });
-      fs.unlinkSync(INFORMATION.image);
-      return true;
+      if (req.role !== null && req.role !== undefined) {
+        const form = await formModel.findOne({ _id: args._id });
+        return form;
+      }
+    } catch (error) {
+      console.log(error);
+      return error;
+    }
+  },
+  deleteForm: async ({ INFORMATION }, req) => {
+    try {
+      if (req.role !== null && req.role !== undefined) {
+        INFORMATION = JSON.parse(JSON.stringify(INFORMATION));
+        await formModel.deleteOne({
+          title: INFORMATION.title,
+          image: INFORMATION.image,
+        });
+        fs.unlinkSync(INFORMATION.image);
+        return true;
+      }
     } catch (error) {
       return false;
     }
